【摘要】mas相关提交

+++ alarm-strategy-service/src/main/java/com/iv/strategy/dao/StrategyPaging.java
@@ -6,12 +6,12 @@ import com.iv.strategy.entity.AlarmStrategyEntity;
 
 
 public class StrategyPaging {
-	int count;
+	long count;
 	List<AlarmStrategyEntity> strategies;
-	public int getCount() {
+	public long getCount() {
 		return count;
 	}
-	public void setCount(int count) {
+	public void setCount(long count) {
 		this.count = count;
 	}
 	public List<AlarmStrategyEntity> getStrategies() {
